Move new relation test app
