+++ tests/integration/conftest.py
@@ -76,3 +76,18 @@ def ops_test(ops_test: OpsTest) -> OpsTest:
 
         ops_test.build_charm = build_charm
     return ops_test
+
+
+@pytest.fixture(scope="module")
+async def postgrsql_test_app_charm(ops_test: OpsTest):
+    """Build the application charm."""
+    charm_path = "tests/integration/postgresql-test-app"
+    charm = await ops_test.build_charm(charm_path)
+    return charm
+
+
+@pytest.fixture(scope="module")
+async def database_charm(ops_test: OpsTest):
+    """Build the database charm."""
+    charm = await ops_test.build_charm(".")
+    return charm
