+++ tests/integration/new_relations/test_new_relations.py
@@ -26,7 +26,7 @@ from tests.integration.new_relations.helpers import (
 
 logger = logging.getLogger(__name__)
 
-APPLICATION_APP_NAME = "application"
+APPLICATION_APP_NAME = "postgresql-test-app"
 DATABASE_APP_NAME = "database"
 ANOTHER_DATABASE_APP_NAME = "another-database"
 DATA_INTEGRATOR_APP_NAME = "data-integrator"
@@ -41,7 +41,7 @@ NO_DATABASE_RELATION_NAME = "no-database"
 
 @pytest.mark.abort_on_fail
 async def test_database_relation_with_charm_libraries(
-    ops_test: OpsTest, application_charm, database_charm
+    ops_test: OpsTest, postgrsql_test_app_charm, database_charm
 ):
     """Test basic functionality of database relation interface."""
     # Deploy both charms (multiple units for each application to test that later they correctly
@@ -49,7 +49,7 @@ async def test_database_relation_with_charm_libraries(
     async with ops_test.fast_forward():
         await asyncio.gather(
             ops_test.model.deploy(
-                application_charm,
+                postgrsql_test_app_charm,
                 application_name=APPLICATION_APP_NAME,
                 num_units=2,
                 series=CHARM_SERIES,
@@ -151,7 +151,7 @@ async def test_user_with_extra_roles(ops_test: OpsTest):
 
 
 async def test_two_applications_doesnt_share_the_same_relation_data(
-    ops_test: OpsTest, application_charm
+    ops_test: OpsTest, postgrsql_test_app_charm
 ):
     """Test that two different application connect to the database with different credentials."""
     # Set some variables to use in this test.
@@ -161,7 +161,7 @@ async def test_two_applications_doesnt_share_the_same_relation_data(
 
     # Deploy another application.
     await ops_test.model.deploy(
-        application_charm,
+        postgrsql_test_app_charm,
         application_name=another_application_app_name,
         series=CHARM_SERIES,
     )
@@ -273,7 +273,7 @@ async def test_an_application_can_connect_to_multiple_aliased_database_clusters(
     assert application_connection_string != another_application_connection_string
 
 
-async def test_an_application_can_request_multiple_databases(ops_test: OpsTest, application_charm):
+async def test_an_application_can_request_multiple_databases(ops_test: OpsTest):
     """Test that an application can request additional databases using the same interface."""
     # Relate the charms using another relation and wait for them exchanging some connection data.
     await ops_test.model.add_relation(
