+++ tests/integration/postgresql-test-app/metadata.yaml
@@ -1,8 +1,8 @@
 # Copyright 2022 Canonical Ltd.
 # See LICENSE file for licensing details.
-name: application
+name: postgresql-test-app
 description: |
-  Data platform libs application charm used in integration tests.
+  Postgresql application charm used in integration tests.
 summary: |
   Data platform libs application meant to be used
   only for testing of the libs in this repository.
